refactor: DTO 클래스의 request/response 폴더 구분
