+++ src/main/java/controller/dto/response/InitialCardStatus.java
@@ -1,4 +1,4 @@
-package controller.dto;
+package controller.dto.response;
 
 import java.util.List;
