+++ src/main/java/controller/dto/response/ParticipantProfitResponse.java
@@ -1,4 +1,4 @@
-package controller.dto;
+package controller.dto.response;
 
 public record ParticipantProfitResponse(
         String name,
