+++ src/main/java/domain/game/ActionAfterPick.java
@@ -1,6 +1,6 @@
 package domain.game;
 
-import controller.dto.ParticipantHandStatus;
+import controller.dto.response.ParticipantHandStatus;
 import java.util.function.Consumer;
 
 public interface ActionAfterPick extends Consumer<ParticipantHandStatus> {
