+++ src/main/java/domain/participant/Participants.java
@@ -1,10 +1,14 @@
 package domain.participant;
 
-import controller.dto.PlayerOutcome;
+import static java.util.stream.Collectors.toList;
+
+import controller.dto.request.PlayerBettingMoney;
+import controller.dto.response.PlayerOutcome;
 import domain.game.deck.PlayerOutcomeFunction;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class Participants {
     private final List<Participant> participants;
@@ -13,19 +17,22 @@ public class Participants {
         this.participants = participants;
     }
 
-    public static Participants from(final List<String> playerNames) {
-        List<Participant> participants = new ArrayList<>();
-        for (String playerName : playerNames) {
-            participants.add(new Player(playerName));
-        }
-        participants.add(new Dealer());
-        return new Participants(participants);
+    public static Participants from(final List<PlayerBettingMoney> requests) {
+        return Stream.concat(
+                Stream.of(new Dealer()),
+                generatePlayers(requests)
+        ).collect(Collectors.collectingAndThen(toList(), Participants::new));
+    }
+
+    private static Stream<Player> generatePlayers(final List<PlayerBettingMoney> requests) {
+        return requests.stream()
+                .map(request -> new Player(request.name(), request.bettingAmount()));
     }
 
     public List<PlayerOutcome> getPlayersOutcomeIf(final PlayerOutcomeFunction function) {
         return getPlayers().stream()
                 .map(player -> new PlayerOutcome(
-                        player.getName(),
+                        player, // TODO: 어짜피 밖에서 NAME 조회해야해서 성능 상 그냥 Player를 바로 보냄.ㄱㅊ?
                         function.apply(player)
                 ))
                 .toList();
